@@ -1,6 +1,6 @@
 {
   "name": "storybook",
-  "version": "5.0.0-beta.4",
+  "version": "5.0.0-beta.5",
   "description": "Storybook components",
   "author": "MUI Team",
   "private": true,
@@ -19,29 +19,29 @@
     "@mui/icons-material": "^5.0.4",
     "@mui/material": "^5.0.4",
     "@mui/styles": "^5.0.1",
-    "@mui/x-data-grid": "5.0.0-beta.4",
-    "@mui/x-data-grid-generator": "5.0.0-beta.4",
-    "@mui/x-data-grid-pro": "5.0.0-beta.4",
-    "@mui/x-license-pro": "5.0.0-beta.4",
-    "@storybook/builder-webpack5": "^6.4.0-beta.13",
-    "@storybook/manager-webpack5": "^6.4.0-beta.13",
+    "@mui/x-data-grid": "5.0.0-beta.5",
+    "@mui/x-data-grid-generator": "5.0.0-beta.5",
+    "@mui/x-data-grid-pro": "5.0.0-beta.5",
+    "@mui/x-license-pro": "5.0.0-beta.5",
+    "@storybook/builder-webpack5": "^6.4.0-beta.19",
+    "@storybook/manager-webpack5": "^6.4.0-beta.19",
     "react": "^17.0.2",
     "react-is": "^17.0.2",
     "rxjs": "^7.4.0",
-    "webpack": "^5.58.2"
+    "webpack": "^5.60.0"
   },
   "devDependencies": {
     "@babel/core": "^7.15.8",
-    "@storybook/addon-a11y": "^6.4.0-beta.13",
-    "@storybook/addon-actions": "^6.4.0-beta.13",
+    "@storybook/addon-a11y": "^6.4.0-beta.19",
+    "@storybook/addon-actions": "^6.4.0-beta.19",
     "@storybook/addon-console": "^1.2.3",
-    "@storybook/addon-docs": "^6.4.0-beta.13",
-    "@storybook/addon-essentials": "^6.4.0-beta.13",
-    "@storybook/addon-storysource": "^6.4.0-beta.13",
-    "@storybook/addon-viewport": "^6.4.0-beta.13",
-    "@storybook/react": "^6.4.0-beta.13",
-    "@storybook/theming": "^6.4.0-beta.13",
-    "babel-loader": "^8.2.2",
+    "@storybook/addon-docs": "^6.4.0-beta.19",
+    "@storybook/addon-essentials": "^6.4.0-beta.19",
+    "@storybook/addon-storysource": "^6.4.0-beta.19",
+    "@storybook/addon-viewport": "^6.4.0-beta.19",
+    "@storybook/react": "^6.4.0-beta.19",
+    "@storybook/theming": "^6.4.0-beta.19",
+    "babel-loader": "^8.2.3",
     "core-js": "^3.6.5",
     "react-docgen-typescript-loader": "^3.7.2",
     "source-map-loader": "^1.1.3",
