@@ -239,55 +239,7 @@ GridColumnHeaderItem.propTypes = {
   // | To update them edit the TypeScript types and run "yarn proptypes"  |
   // ----------------------------------------------------------------------
   colIndex: PropTypes.number.isRequired,
-  column: PropTypes.shape({
-    align: PropTypes.oneOf(['center', 'left', 'right']),
-    cellClassName: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-    computedWidth: PropTypes.number.isRequired,
-    description: PropTypes.string,
-    disableColumnMenu: PropTypes.bool,
-    disableExport: PropTypes.bool,
-    disableReorder: PropTypes.bool,
-    editable: PropTypes.bool,
-    field: PropTypes.string.isRequired,
-    filterable: PropTypes.bool,
-    filterOperators: PropTypes.arrayOf(
-      PropTypes.shape({
-        getApplyFilterFn: PropTypes.func.isRequired,
-        InputComponent: PropTypes.elementType,
-        InputComponentProps: PropTypes.object,
-        label: PropTypes.string,
-        value: PropTypes.string.isRequired,
-      }),
-    ),
-    flex: PropTypes.number,
-    headerAlign: PropTypes.oneOf(['center', 'left', 'right']),
-    headerClassName: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-    headerName: PropTypes.string,
-    hide: PropTypes.bool,
-    hideSortIcons: PropTypes.bool,
-    minWidth: PropTypes.number,
-    renderCell: PropTypes.func,
-    renderEditCell: PropTypes.func,
-    renderHeader: PropTypes.func,
-    resizable: PropTypes.bool,
-    sortable: PropTypes.bool,
-    sortComparator: PropTypes.func,
-    type: PropTypes.string,
-    valueFormatter: PropTypes.func,
-    valueGetter: PropTypes.func,
-    valueOptions: PropTypes.arrayOf(
-      PropTypes.oneOfType([
-        PropTypes.number,
-        PropTypes.shape({
-          label: PropTypes.string.isRequired,
-          value: PropTypes.any.isRequired,
-        }),
-        PropTypes.string,
-      ]).isRequired,
-    ),
-    valueParser: PropTypes.func,
-    width: PropTypes.number,
-  }).isRequired,
+  column: PropTypes.object.isRequired,
   columnMenuOpen: PropTypes.bool.isRequired,
   extendRowFullWidth: PropTypes.bool.isRequired,
   filterItemsCounter: PropTypes.number,
