@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import { useForkRef } from '@mui/material/utils';
 import NoSsr from '@mui/material/NoSsr';
 import { GridRootContainerRef } from '../../models/gridRootContainerRef';
-import { useStyles } from './GridRootStyles';
+import { GridRootStyles } from './GridRootStyles';
 import { visibleGridColumnsLengthSelector } from '../../hooks/features/columns/gridColumnsSelector';
 import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { useGridApiContext } from '../../hooks/utils/useGridApiContext';
@@ -17,10 +17,9 @@ export const GridRoot = React.forwardRef<HTMLDivElement, GridRootProps>(function
   props,
   ref,
 ) {
-  const stylesClasses = useStyles();
-  const apiRef = useGridApiContext();
   const rootProps = useGridRootProps();
-  const { children, className: classNameProp, ...other } = props;
+  const { children, className, ...other } = props;
+  const apiRef = useGridApiContext();
   const visibleColumnsLength = useGridSelector(apiRef, visibleGridColumnsLengthSelector);
   const totalRowCount = useGridSelector(apiRef, gridRowCountSelector);
   const rootContainerRef: GridRootContainerRef = React.useRef<HTMLDivElement>(null);
@@ -30,18 +29,11 @@ export const GridRoot = React.forwardRef<HTMLDivElement, GridRootProps>(function
 
   return (
     <NoSsr>
-      <div
+      <GridRootStyles
         ref={handleRef}
-        className={clsx(
-          stylesClasses.root,
-          rootProps.classes?.root,
-          rootProps.className,
-          classNameProp,
-          gridClasses.root,
-          {
-            [gridClasses.autoHeight]: rootProps.autoHeight,
-          },
-        )}
+        className={clsx(className, rootProps.classes?.root, rootProps.className, gridClasses.root, {
+          [gridClasses.autoHeight]: rootProps.autoHeight,
+        })}
         role="grid"
         aria-colcount={visibleColumnsLength}
         aria-rowcount={totalRowCount}
@@ -52,7 +44,7 @@ export const GridRoot = React.forwardRef<HTMLDivElement, GridRootProps>(function
         {...other}
       >
         {children}
-      </div>
+      </GridRootStyles>
     </NoSsr>
   );
 });
