@@ -1,35 +1,48 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { makeStyles } from '@mui/styles';
-import { createTheme } from '@mui/material/styles';
-
-const defaultTheme = createTheme();
-const useStyles = makeStyles(
-  {
-    root: {
-      display: 'flex',
-      alignItems: 'center',
-      padding: '4px 4px 0',
-    },
-  },
-  { name: 'MuiGridToolbarContainer', defaultTheme },
-);
+import { styled } from '@mui/material/styles';
+import { unstable_composeClasses as composeClasses } from '@mui/material';
+import { GridComponentProps } from '../../GridComponentProps';
+import { getDataGridUtilityClass } from '../../gridClasses';
+import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
 
 export type GridToolbarContainerProps = React.HTMLAttributes<HTMLDivElement>;
+
+type OwnerState = { classes: GridComponentProps['classes'] };
+
+const useUtilityClasses = (ownerState: OwnerState) => {
+  const { classes } = ownerState;
+
+  const slots = {
+    root: ['toolbarContainer'],
+  };
+
+  return composeClasses(slots, getDataGridUtilityClass, classes);
+};
+
+const GridToolbarContainerRoot = styled('div', {
+  name: 'MuiDataGrid',
+  slot: 'toolbarContainer',
+})(({ theme }) => ({
+  display: 'flex',
+  alignItems: 'center',
+  padding: theme.spacing(0.5, 0.5, 0),
+}));
 
 export const GridToolbarContainer = React.forwardRef<HTMLDivElement, GridToolbarContainerProps>(
   function GridToolbarContainer(props, ref) {
     const { className, children, ...other } = props;
-    const classes = useStyles();
-
+    const rootProps = useGridRootProps();
+    const ownerState = { classes: rootProps.classes };
+    const classes = useUtilityClasses(ownerState);
     if (!children) {
       return null;
     }
 
     return (
-      <div ref={ref} className={clsx(classes.root, className)} {...other}>
+      <GridToolbarContainerRoot ref={ref} className={clsx(className, classes.root)} {...other}>
         {children}
-      </div>
+      </GridToolbarContainerRoot>
     );
   },
 );
