@@ -1,12 +1,16 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import clsx from 'clsx';
 import ClickAwayListener, { ClickAwayListenerProps } from '@mui/material/ClickAwayListener';
+import { unstable_composeClasses as composeClasses } from '@mui/material';
 import Grow from '@mui/material/Grow';
 import Paper from '@mui/material/Paper';
 import Popper, { PopperProps } from '@mui/material/Popper';
-import { createTheme } from '@mui/material/styles';
-import { makeStyles } from '@mui/styles';
+import { styled } from '@mui/material/styles';
 import { HTMLElementType } from '@mui/utils';
+import { getDataGridUtilityClass, gridClasses } from '../../gridClasses';
+import { GridComponentProps } from '../../GridComponentProps';
+import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
 
 type MenuPosition =
   | 'bottom-end'
@@ -23,19 +27,27 @@ type MenuPosition =
   | 'top'
   | undefined;
 
-const defaultTheme = createTheme();
-const useStyles = makeStyles(
-  (theme) => ({
-    root: {
-      zIndex: theme.zIndex.modal,
-      '& .MuiDataGrid-gridMenuList': {
-        // TODO: Rename to MuiGridMenu-menuList
-        outline: 0,
-      },
-    },
-  }),
-  { name: 'MuiGridMenu', defaultTheme },
-);
+type OwnerState = { classes: GridComponentProps['classes'] };
+
+const useUtilityClasses = (ownerState: OwnerState) => {
+  const { classes } = ownerState;
+
+  const slots = {
+    root: ['menu'],
+  };
+
+  return composeClasses(slots, getDataGridUtilityClass, classes);
+};
+
+const StyledPopper = styled(Popper, {
+  name: 'MuiDataGrid',
+  slot: 'Menu',
+})(({ theme }) => ({
+  zIndex: theme.zIndex.modal,
+  [`& .${gridClasses.menuList}`]: {
+    outline: 0,
+  },
+}));
 
 export interface GridMenuProps extends Omit<PopperProps, 'onKeyDown'> {
   open: boolean;
@@ -50,10 +62,12 @@ const transformOrigin = {
 };
 
 const GridMenu = (props: GridMenuProps) => {
-  const { open, target, onClickAway, children, position, ...other } = props;
+  const { open, target, onClickAway, children, position, className, ...other } = props;
   const prevTarget = React.useRef(target);
   const prevOpen = React.useRef(open);
-  const classes = useStyles();
+  const rootProps = useGridRootProps();
+  const ownerState = { classes: rootProps.classes };
+  const classes = useUtilityClasses(ownerState);
 
   React.useEffect(() => {
     if (prevOpen.current && prevTarget.current) {
@@ -65,8 +79,8 @@ const GridMenu = (props: GridMenuProps) => {
   }, [open, target]);
 
   return (
-    <Popper
-      className={classes.root}
+    <StyledPopper
+      className={clsx(className, classes.root)}
       open={open}
       anchorEl={target as any}
       transition
@@ -80,7 +94,7 @@ const GridMenu = (props: GridMenuProps) => {
           </Grow>
         </ClickAwayListener>
       )}
-    </Popper>
+    </StyledPopper>
   );
 };
 
