@@ -1,10 +1,11 @@
 import * as React from 'react';
+import { unstable_composeClasses as composeClasses } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
-import Switch from '@mui/material/Switch';
+import Switch, { switchClasses } from '@mui/material/Switch';
 import Button from '@mui/material/Button';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import TextField from '@mui/material/TextField';
-import { makeStyles } from '@mui/styles';
+import { styled } from '@mui/material/styles';
 import { allGridColumnsSelector } from '../../hooks/features/columns/gridColumnsSelector';
 import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { useGridApiContext } from '../../hooks/utils/useGridApiContext';
@@ -15,34 +16,53 @@ import { GridPanelHeader } from './GridPanelHeader';
 import { GridPanelWrapper } from './GridPanelWrapper';
 import { GRID_EXPERIMENTAL_ENABLED } from '../../constants';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
+import { GridComponentProps } from '../../GridComponentProps';
+import { getDataGridUtilityClass } from '../../gridClasses';
 
-const useStyles = makeStyles(
-  {
-    container: {
-      padding: '8px 0px 8px 8px',
-    },
-    column: {
-      display: 'flex',
-      justifyContent: 'space-between',
-      padding: '1px 8px 1px 7px',
-    },
-    switch: {
-      marginRight: 4,
-    },
-    dragIcon: {
-      justifyContent: 'flex-end',
-    },
+type OwnerState = { classes: GridComponentProps['classes'] };
+
+const useUtilityClasses = (ownerState: OwnerState) => {
+  const { classes } = ownerState;
+
+  const slots = {
+    root: ['columnsPanel'],
+    columnsPanelRow: ['columnsPanelRow'],
+  };
+
+  return composeClasses(slots, getDataGridUtilityClass, classes);
+};
+
+const GridColumnsPanelRoot = styled('div', {
+  name: 'MuiDataGrid',
+  slot: 'ColumnsPanel',
+})(() => ({
+  padding: '8px 0px 8px 8px',
+}));
+
+const GridColumnsPanelRowRoot = styled('div', {
+  name: 'MuiDataGrid',
+  slot: 'ColumnsPanelRow',
+})(({ theme }) => ({
+  display: 'flex',
+  justifyContent: 'space-between',
+  padding: '1px 8px 1px 7px',
+  [`& .${switchClasses.root}`]: {
+    marginRight: theme.spacing(0.5),
   },
-  { name: 'MuiDataGridColumnsPanel' }, // TODO rename to MuiGridColumnsPanel
-);
+}));
+
+const StyledIconButton = styled(IconButton)({
+  justifyContent: 'flex-end',
+});
 
 export function GridColumnsPanel() {
-  const classes = useStyles();
   const apiRef = useGridApiContext();
   const searchInputRef = React.useRef<HTMLInputElement>(null);
   const columns = useGridSelector(apiRef, allGridColumnsSelector);
   const rootProps = useGridRootProps();
   const [searchValue, setSearchValue] = React.useState('');
+  const ownerState = { classes: rootProps.classes };
+  const classes = useUtilityClasses(ownerState);
 
   const toggleColumn = React.useCallback(
     (event: React.MouseEvent<HTMLButtonElement>) => {
@@ -64,6 +84,7 @@ export function GridColumnsPanel() {
     },
     [apiRef, columns],
   );
+
   const showAllColumns = React.useCallback(() => toggleAllColumns(false), [toggleAllColumns]);
   const hideAllColumns = React.useCallback(() => toggleAllColumns(true), [toggleAllColumns]);
 
@@ -102,13 +123,12 @@ export function GridColumnsPanel() {
         />
       </GridPanelHeader>
       <GridPanelContent>
-        <div className={classes.container}>
+        <GridColumnsPanelRoot className={classes.root}>
           {currentColumns.map((column) => (
-            <div key={column.field} className={classes.column}>
+            <GridColumnsPanelRowRoot className={classes.columnsPanelRow} key={column.field}>
               <FormControlLabel
                 control={
                   <Switch
-                    className={classes.switch}
                     checked={!column.hide}
                     onClick={toggleColumn}
                     name={column.field}
@@ -119,20 +139,19 @@ export function GridColumnsPanel() {
                 label={column.headerName || column.field}
               />
               {!rootProps.disableColumnReorder && GRID_EXPERIMENTAL_ENABLED && (
-                <IconButton
+                <StyledIconButton
                   draggable
-                  className={classes.dragIcon}
                   aria-label={apiRef.current.getLocaleText('columnsPanelDragIconLabel')}
                   title={apiRef.current.getLocaleText('columnsPanelDragIconLabel')}
                   size="small"
                   disabled
                 >
                   <GridDragIcon />
-                </IconButton>
+                </StyledIconButton>
               )}
-            </div>
+            </GridColumnsPanelRowRoot>
           ))}
-        </div>
+        </GridColumnsPanelRoot>
       </GridPanelContent>
       <GridPanelFooter>
         <Button onClick={hideAllColumns} color="primary">
