@@ -1,6 +1,4 @@
 import * as React from 'react';
-import clsx from 'clsx';
-import { unstable_composeClasses as composeClasses } from '@mui/material';
 import {
   GridToolbarContainer,
   GridToolbarContainerProps,
@@ -10,27 +8,11 @@ import { GridToolbarDensitySelector } from './GridToolbarDensitySelector';
 import { GridToolbarFilterButton } from './GridToolbarFilterButton';
 import { GridToolbarExport } from './GridToolbarExport';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
-import { GridComponentProps } from '../../GridComponentProps';
-import { getDataGridUtilityClass } from '../../gridClasses';
-
-type OwnerState = { classes: GridComponentProps['classes'] };
-
-const useUtilityClasses = (ownerState: OwnerState) => {
-  const { classes } = ownerState;
-
-  const slots = {
-    root: ['toolbarContainer'],
-  };
-
-  return composeClasses(slots, getDataGridUtilityClass, classes);
-};
 
 export const GridToolbar = React.forwardRef<HTMLDivElement, GridToolbarContainerProps>(
   function GridToolbar(props, ref) {
     const { className, ...other } = props;
     const rootProps = useGridRootProps();
-    const ownerState = { classes: rootProps.classes };
-    const classes = useUtilityClasses(ownerState);
 
     if (
       rootProps.disableColumnFilter &&
@@ -41,7 +23,7 @@ export const GridToolbar = React.forwardRef<HTMLDivElement, GridToolbarContainer
     }
 
     return (
-      <GridToolbarContainer ref={ref} className={clsx(className, classes.root)} {...other}>
+      <GridToolbarContainer ref={ref} {...other}>
         <GridToolbarColumnsButton />
         <GridToolbarFilterButton />
         <GridToolbarDensitySelector />
