@@ -12,6 +12,7 @@ import { useGridSelector } from '../../hooks/utils/useGridSelector';
 import { GridDensityOption } from '../../models/api/gridDensityApi';
 import { GridMenu } from '../menu/GridMenu';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
+import { gridClasses } from '../../gridClasses';
 
 export const GridToolbarDensitySelector = React.forwardRef<HTMLButtonElement, ButtonProps>(
   function GridToolbarDensitySelector(props, ref) {
@@ -115,7 +116,7 @@ export const GridToolbarDensitySelector = React.forwardRef<HTMLButtonElement, Bu
         >
           <MenuList
             id={densityMenuId}
-            className="MuiDataGrid-gridMenuList"
+            className={gridClasses.menuList}
             aria-labelledby={densityButtonId}
             onKeyDown={handleListKeyDown}
             autoFocusItem={open}
