@@ -20,7 +20,7 @@ export interface GridRowGroupsPreProcessingApi {
    * @param {GridRowGroupingPreProcessing} columnsPreProcessing Pre-processing to register.
    * @ignore - do not document
    */
-  UNSTABLE_registerRowGroupsBuilder: (
+  unstable_registerRowGroupsBuilder: (
     processingName: string,
     groupingFunction: GridRowGroupingPreProcessing | null,
   ) => void;
@@ -32,5 +32,5 @@ export interface GridRowGroupsPreProcessingApi {
    * @returns {GridRowGroupingResult} The grouped rows
    * @ignore - do not document
    */
-  UNSTABLE_groupRows: (params: RowGroupParams) => GridRowGroupingResult;
+  unstable_groupRows: (params: RowGroupParams) => GridRowGroupingResult;
 }
