@@ -3,8 +3,7 @@ import { GridApiRef } from '../../../models/api/gridApiRef';
 import { useGridApiMethod } from '../../utils/useGridApiMethod';
 import { useGridSelector } from '../../utils/useGridSelector';
 import { allGridColumnsSelector, visibleGridColumnsSelector } from '../columns';
-import { visibleSortedGridRowsSelector } from '../filter';
-import { gridSelectionStateSelector } from '../selection';
+import { visibleSortedGridRowIdsSelector } from '../filter';
 import { GridCsvExportApi } from '../../../models/api/gridCsvExportApi';
 import { GridCsvExportOptions } from '../../../models/gridExport';
 import { useGridLogger } from '../../utils/useGridLogger';
@@ -22,9 +21,8 @@ import { GridStateColDef } from '../../../models';
 export const useGridCsvExport = (apiRef: GridApiRef): void => {
   const logger = useGridLogger(apiRef, 'useGridCsvExport');
   const visibleColumns = useGridSelector(apiRef, visibleGridColumnsSelector);
+  const visibleSortedRowIds = useGridSelector(apiRef, visibleSortedGridRowIdsSelector);
   const columns = useGridSelector(apiRef, allGridColumnsSelector);
-  const visibleSortedRows = useGridSelector(apiRef, visibleSortedGridRowsSelector);
-  const selection = useGridSelector(apiRef, gridSelectionStateSelector);
 
   const getDataAsCsv = React.useCallback(
     (options?: GridCsvExportOptions): string => {
@@ -38,19 +36,24 @@ export const useGridCsvExport = (apiRef: GridApiRef): void => {
           .filter((column): column is GridStateColDef => !!column);
       } else {
         const validColumns = options?.allColumns ? columns : visibleColumns;
-
         exportedColumns = validColumns.filter((column) => !column.disableExport);
       }
 
+      const selectedRows = apiRef.current.getSelectedRows();
+      const exportedRowIds =
+        selectedRows.size > 0
+          ? visibleSortedRowIds.filter((id) => selectedRows.has(id))
+          : visibleSortedRowIds;
+
       return buildCSV({
         columns: exportedColumns,
-        rows: visibleSortedRows,
-        selectedRowIds: selection,
+        rowIds: exportedRowIds,
         getCellParams: apiRef.current.getCellParams,
         delimiterCharacter: options?.delimiter || ',',
+        includeHeaders: options?.includeHeaders ?? true,
       });
     },
-    [logger, visibleColumns, columns, visibleSortedRows, selection, apiRef],
+    [logger, visibleColumns, columns, visibleSortedRowIds, apiRef],
   );
 
   const exportDataAsCsv = React.useCallback(
