@@ -6,10 +6,10 @@ export interface GridDisableVirtualizationApi {
    * Disables grid's virtualization.
    * @ignore - do not document. Remove before releasing v5 stable version.
    */
-  UNSTABLE_disableVirtualization: () => void;
+  unstable_disableVirtualization: () => void;
   /**
    * Enables grid's virtualization.
    * @ignore - do not document. Remove before releasing v5 stable version.
    */
-  UNSTABLE_enableVirtualization: () => void;
+  unstable_enableVirtualization: () => void;
 }
