@@ -408,7 +408,7 @@ describe('<DataGridPro /> - Filter', () => {
       render(<AddServerFilterGrid />);
       const addButton = screen.getByRole('button', { name: /Add Filter/i });
       fireEvent.click(addButton);
-      const filterForms = document.querySelectorAll(`.MuiGridFilterForm-root`);
+      const filterForms = document.querySelectorAll(`.MuiDataGrid-filterForm`);
       expect(filterForms).to.have.length(2);
     });
   });
@@ -434,7 +434,7 @@ describe('<DataGridPro /> - Filter', () => {
       );
       const addButton = screen.getByRole('button', { name: /Add Filter/i });
       fireEvent.click(addButton);
-      const filterForms = document.querySelectorAll(`.MuiGridFilterForm-root`);
+      const filterForms = document.querySelectorAll(`.MuiDataGrid-filterForm`);
       expect(filterForms).to.have.length(2);
     });
 
@@ -453,7 +453,7 @@ describe('<DataGridPro /> - Filter', () => {
       );
       const addButton = screen.getByRole('button', { name: /Add Filter/i });
       fireEvent.click(addButton);
-      const filterForms = document.querySelectorAll(`.MuiGridFilterForm-root`);
+      const filterForms = document.querySelectorAll(`.MuiDataGrid-filterForm`);
       expect(filterForms).to.have.length(0);
     });
 
@@ -473,7 +473,7 @@ describe('<DataGridPro /> - Filter', () => {
       expect(onModelChange.callCount).to.equal(1);
       const addButton = screen.getByRole('button', { name: /Add Filter/i });
       fireEvent.click(addButton);
-      const filterForms = document.querySelectorAll(`.MuiGridFilterForm-root`);
+      const filterForms = document.querySelectorAll(`.MuiDataGrid-filterForm`);
       expect(filterForms).to.have.length(2);
       expect(onModelChange.callCount).to.equal(2);
       expect(onModelChange.lastCall.firstArg.items.length).to.deep.equal(2);
@@ -512,7 +512,7 @@ describe('<DataGridPro /> - Filter', () => {
       const addButton = screen.getByRole('button', { name: /Add Filter/i });
       fireEvent.click(addButton);
 
-      const filterForms = document.querySelectorAll(`.MuiGridFilterForm-root`);
+      const filterForms = document.querySelectorAll(`.MuiDataGrid-filterForm`);
       expect(filterForms).to.have.length(2);
     });
   });
